requirements.txt
# Broadway MDS-SV Analysis Pipeline
# Python 3.9+ required

# Core data processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# Google Trends
pytrends==4.9.2

# Machine learning (for PCA)
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
tqdm==4.66.1
tenacity==8.2.3

# Optional: Browser automation for dynamic scraping
# playwright==1.40.0

# Jupyter notebooks (optional)
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Type checking (development)
# mypy==1.7.1
